+++ token.h
@@ -8,16 +8,16 @@
 class token {
 private:
 	tokentype type;
-	double data;
+	long double data;
 	std::string name;
 
 public:
-	token(tokentype typ, double dat);
+	token(tokentype typ, long double dat);
 	token(tokentype typ, std::string nam);
 	token(tokentype typ);
 
 	tokentype get_type();
-	double get_double_value();
+	long double get_double_value();
 	std::string get_string_value();
 
 	friend std::ostream& operator<<(std::ostream& stream, token& tok);
